+++ internal/impl/pure/input_read_until_test.go
@@ -268,3 +268,28 @@ remainingLoop:
 		t.Fatal(err)
 	}
 }
+
+func TestReadUntilTimeout(t *testing.T) {
+	conf := input.NewConfig()
+	require.NoError(t, yaml.Unmarshal([]byte(`
+read_until:
+  idle_timeout: 100ms
+  input:
+    generate:
+      count: 1000
+      interval: 1s
+      mapping: 'root.id = counter()'
+`), &conf))
+
+	strm, err := bmock.NewManager().NewInput(conf)
+	require.NoError(t, err)
+
+	tran, open := <-strm.TransactionChan()
+	require.True(t, open)
+	require.Len(t, tran.Payload, 1)
+	assert.Equal(t, `{"id":1}`, string(tran.Payload[0].AsBytes()))
+	require.NoError(t, tran.Ack(context.Background(), nil))
+
+	_, open = <-strm.TransactionChan()
+	require.False(t, open)
+}